# Conflicts:
#	app/src/main/java/com/pes/become/backend/domain/Achievement.java
#	app/src/main/java/com/pes/become/backend/domain/User.java

+++ app/src/main/java/com/pes/become/backend/persistence/ControllerPersistence.java
@@ -2,22 +2,51 @@ package com.pes.become.backend.persistence;
 
 import android.app.Activity;
 import android.net.Uri;
-import android.util.Log;
+
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.lang.reflect.Method;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.zip.DeflaterOutputStream;
 
 public class ControllerPersistence {
 
-    ControllerRoutineDB CR;
-    ControllerActivityDB CA;
-    ControllerUserDB CU;
+    /**
+     * Unica instancia de la classe
+     */
+    private static ControllerPersistence instance;
+    private final ControllerRoutineDB CR;
+    private final ControllerActivityDB CA;
+    private final ControllerUserDB CU;
+    private final ControllerStatisticsDB CS;
+    private final ControllerCalendarDB CD;
+    private final ControllerTrophiesDB CT;
+    private final FirebaseFirestore db;
+
+    /**
+     * Obtenir la instancia de la classe
+     * @return instancia
+     */
+    public static ControllerPersistence getInstance(){
+        if(instance == null)
+            instance = new ControllerPersistence();
+        return instance;
+    }
 
     /**
      * Creadora per defecte de la classe ControllerPersistence
      */
-    public ControllerPersistence() {
-        CA = new ControllerActivityDB();
-        CR = new ControllerRoutineDB();
+    private ControllerPersistence() {
+        CA = ControllerActivityDB.getInstance();
+        CR = ControllerRoutineDB.getInstance();
+        CU = ControllerUserDB.getInstance();
+        CS = ControllerStatisticsDB.getInstance();
+        CD = ControllerCalendarDB.getInstance();
+        CT = ControllerTrophiesDB.getInstance();
+        db = FirebaseFirestore.getInstance();
     }
 
     /**
@@ -29,18 +58,6 @@ public class ControllerPersistence {
      */
     public void getActivitiesRoutine(String userId, String idRoutine, Method method, Object object) {
         CA.getActivities(userId, idRoutine, method, object);
-    }
-
-    /**
-     * Obtenir les activitats d'una rutina i un dia indicats
-     * @param userId identificador de l'usuari
-     * @param idRoutine identificador de la rutina
-     * @param day dia a consultar
-     * @param method metode a cridar quan es retornin les dades
-     * @param object classe que conté el mètode
-     */
-    public void getActivitiesByDay(String userId, String idRoutine, String day, Method method, Object object) {
-        CA.getActivitiesByDay(userId, idRoutine,day,method,object);
     }
 
     /**
@@ -56,33 +73,82 @@ public class ControllerPersistence {
      * @return el valor del id de l'activitat creada
      */
     public String createActivity(String userId, String idRoutine, String activityName, String actTheme,String actDescription, String actDay, String beginTime, String finishTime) {
-        return CA.createActivity(userId, idRoutine, activityName,actTheme,actDescription, actDay, beginTime, finishTime);
+        CS.addActivityToStatistics(userId, idRoutine, actTheme, actDay, beginTime, finishTime);
+        return CA.createActivity(userId, idRoutine, activityName,actTheme,actDescription, actDay, beginTime, finishTime, "null");
     }
 
     /**
      * Esborrar una activitat d'una rutina
      * @param userId identificador de l'usuari
-     * @param idRoutine és el nom i l'identificador de la rutina
-     * @param idActivity és l'identificador de l'activitat
+     * @param idRoutine identificador de la rutina
+     * @param idActivity identificador de l'activitat
      */
     public void deleteActivity(String userId, String idRoutine, String idActivity) {
-        CA.deleteActivity(userId, idRoutine, idActivity);
+
+        DocumentReference docRefToActivity = db.collection("users").document(userId).
+                collection("routines").document(idRoutine).
+                collection("activities").document(idActivity);
+
+        docRefToActivity.get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                DocumentSnapshot document = task.getResult();
+                if (document.exists()) {
+                    String day = document.get("day").toString();
+                    String theme = document.get("theme").toString();
+                    String beginTime = document.get("beginTime").toString();
+                    String finishTime = document.get("finishTime").toString();
+                    CS.deleteActivityStatistics(userId, idRoutine, day, theme, beginTime, finishTime);
+                    CA.deleteActivity(userId, idRoutine, idActivity);
+
+                }
+            }
+        });
     }
 
     /**
      * Actualitzar una activitat existent en una rutina existent
      * @param userId identificador de l'usuari
-     * @param idRoutine és l'identificador de la rutina ja existent
-     * @param actName és el nom de l'activitat que es vol modificar
-     * @param description és la nova descripció que es vol afegir a l'activitat
-     * @param theme és el tema que es vol afegir a l'activitat
-     * @param day és el dia de  l'activitat que es vol modificar
-     * @param iniT és l'hora d'inici de l'activitat
-     * @param endT és l'hora d'acabament de l'activitat
-     * @param idActivity és l'identificador de l'activitat
+     * @param idRoutine identificador de la rutina
+     * @param idActivity identificador de l'activitat
+     * @param actName nou nom de l'activitat
+     * @param description nova descripcio de l'activitat
+     * @param newDay nou dia de l'activitat
+     * @param newTheme nou tema de l'activitat
+     * @param newBeginTime nova hora d'inici de l'activitat
+     * @param newFinishTime nova hora final de l'activitat
      */
-    public void updateActivity(String userId, String idRoutine, String actName, String description, String theme,  String iniT, String endT, String day,String idActivity) {
-        CA.updateActivity(userId, idRoutine, actName, description, theme, day, iniT, endT, idActivity);
+    public void updateActivity(String userId, String idRoutine, String idActivity, String actName, String description, String newDay, String newTheme,  String newBeginTime, String newFinishTime) {
+
+        DocumentReference docRefToActivity = db.collection("users").document(userId).
+                collection("routines").document(idRoutine).
+                collection("activities").document(idActivity);
+
+        docRefToActivity.get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                DocumentSnapshot document = task.getResult();
+                if (document.exists()) {
+                    String oldDay = document.get("day").toString();
+                    String oldTheme = document.get("theme").toString();
+                    String oldBeginTime = document.get("beginTime").toString();
+                    String oldFinishTime = document.get("finishTime").toString();
+                    CS.updateDedicatedTimeActivity(userId, idRoutine, newDay, newTheme, newBeginTime, newFinishTime, oldDay, oldTheme, oldBeginTime, oldFinishTime);
+                    CA.updateActivity(userId, idRoutine, actName, description, newTheme, newDay, newBeginTime, newFinishTime, idActivity);
+
+                }
+            }
+        });
+    }
+
+    /** Marca una activitat com a feta i actualitza el calendari.
+     * @param userId identificador de l'usuari
+     * @param idRoutine es l'identificador de la rutina ja existent
+     * @param lastDayDone és l'últim dia que l'usuari ha marcat la rutina com a feta en format de data yyyy-mm-dd (la classe StringDateConverter serveix per convertir-la)
+     * @param idActivity és l'identificador de l'activitat
+     * @param totalActivities nombre total d'activitats del dia per la rutina seleccionada
+     */
+    public void markActivityAsDone(String userId, String idRoutine, String lastDayDone, String idActivity, int totalActivities)
+    {
+        CA.markActivityAsDone(userId, idRoutine, lastDayDone, idActivity, totalActivities);
     }
 
     /**
@@ -101,7 +167,9 @@ public class ControllerPersistence {
      * @param routineName és el nom que es vol que tingui la rutina
      */
     public String createRoutine(String userId, String routineName) {
-        return CR.createRoutine(userId, routineName);
+        String idRoutine = CR.createRoutine(userId, routineName);
+        CS.createRoutineStatistics(userId,idRoutine);
+        return idRoutine;
     }
 
     /**
@@ -111,6 +179,7 @@ public class ControllerPersistence {
      */
     public void deleteRoutine(String userId, String idRoutine) {
         CR.deleteRoutine(userId, idRoutine);
+        CS.deleteRoutineStatistics(userId, idRoutine);
     }
 
     /**
@@ -118,8 +187,16 @@ public class ControllerPersistence {
      * @return el provider de l'usuari
      */
     public String getUserProvider() {
-        CU = ControllerUserDB.getInstance();
         return CU.getUserProvider();
+    }
+
+    /**
+     * Metode per carregar un usuari
+     * @param method metode a cridar quan es retornin les dades
+     * @param object classe que conté el mètode
+     */
+    public void loadUser(Method method, Object object) {
+        CU.loadUser(method, object);
     }
 
     /**
@@ -128,9 +205,8 @@ public class ControllerPersistence {
      * @param method mètode a executar de forma asíncrona un cop acabada la reautentificació (el paràmetre és un boolea que retorna true si la reautentificació ha anat bé o false si no)
      * @param object instancia de la classe del mètode a executar
      */
-    public void changePassword(String newPassword, Method method, Object object) {
-        CU = ControllerUserDB.getInstance();
-        CU.changePassword(newPassword, method, object);
+    public void changePassword(String oldPassword, String newPassword, Method method, Object object) {
+        CU.changePassword(oldPassword, newPassword, method, object);
     }
 
     /**
@@ -139,7 +215,6 @@ public class ControllerPersistence {
      * @param imageUri uri de la imatge a penjar
      */
     public void updateProfilePic(String userId, Uri imageUri) {
-        CU = ControllerUserDB.getInstance();
         CU.updateProfilePic(userId, imageUri);
     }
 
@@ -149,7 +224,6 @@ public class ControllerPersistence {
      * @param object instancia de la classe del mètode a executar
      */
     public void deleteUser(String password, Method method, Object object) {
-        CU = ControllerUserDB.getInstance();
         CU.deleteUser(password, method, object);
     }
 
@@ -162,7 +236,6 @@ public class ControllerPersistence {
      * @param object classe que conté el mètode
      */
     public void loginUser(String mail, String password, Activity act, Method method, Object object) {
-        CU = ControllerUserDB.getInstance();
         CU.loginUser(mail, password, act, method, object);
     }
 
@@ -173,7 +246,6 @@ public class ControllerPersistence {
      * @param object classe que conté el mètode
      */
     public void loginUserGoogle(String idToken, Method method, Object object ){
-        CU = ControllerUserDB.getInstance();
         CU.loginUserGoogle(idToken, method, object);
     }
 
@@ -187,7 +259,6 @@ public class ControllerPersistence {
      * @param object classe que conté el mètode
      */
     public void registerUser(String mail, String password, String name, Activity act, Method method, Object object) {
-        CU = ControllerUserDB.getInstance();
         CU.registerUser(mail, password, name, act, method, object);
     }
 
@@ -197,7 +268,6 @@ public class ControllerPersistence {
      * @param routineID identificador de la nova rutina seleccionada
      */
     public void setSelectedRoutine(String userID, String routineID) {
-        CU = ControllerUserDB.getInstance();
         CU.setSelectedRoutine(userID, routineID);
     }
 
@@ -205,18 +275,7 @@ public class ControllerPersistence {
      * Metode per tancar la sessió
      */
     public void signOut() {
-        CU = ControllerUserDB.getInstance();
         CU.signOut();
-    }
-
-    /**
-     * Metode per carregar un usuari
-     * @param method metode a cridar quan es retornin les dades
-     * @param object classe que conté el mètode
-     */
-    public void loadUser(Method method, Object object) {
-        CU = ControllerUserDB.getInstance();
-        CU.loadUser(method, object);
     }
 
     /**
@@ -224,8 +283,127 @@ public class ControllerPersistence {
      * @param mail mail del compte a recuperar
      */
     public void sendPassResetEmail(String mail, Method method, Object object) {
-        CU = ControllerUserDB.getInstance();
         CU.sendPassResetEmail(mail, method, object);
+    }
+
+    /**
+     * Metode per canviar el nom de l'usuari
+     * @param newName nou nom de l'usuari
+     */
+    public void changeUsername(String userID, String newName) {
+        CU.changeUsername(userID, newName);
+    }
+
+    /**
+     * Funcio per aconseguir totes les estadistiques d'una rutina
+     * @param userId identificador de l'usuari
+     * @param idRoutine identificador de la rutina
+     * @param method metode a cridar per retornar les dades
+     * @param object classe que conte el metode
+     */
+    public void getAllStatisticsRoutine(String userId, String idRoutine, Method method, Object object){
+        CS.getAllStatisticsRoutine(userId, idRoutine, method, object);
+    }
+
+    /**
+     * Funcio per aconseguir les estadistiques d'una rutina i d'un tema concret
+     * @param userId identificador de l'usuari
+     * @param idRoutine identificador de la rutina
+     * @param theme tema del que es vol aconseguir les estadistiques
+     * @param method metode a cridar per retornar les dades
+     * @param object classe que conte el metode
+     */
+    public void getStatisticsRoutineByTheme(String userId, String idRoutine, String theme, Method method, Object object){
+        CS.getStatisticsRoutineByTheme(userId, idRoutine, theme, method, object);
+    }
+
+    /**
+     * Executa el metode method amb un hashmap que representa el day de la base de dades si aquest s'ha pogut consultar, o l'excepció que ha saltat si no.
+     * Day tindrà les claus: day, idRoutine, numActivitiesDone, numTotalActivities. Totes son strings
+     * @param userId id de l'usuari del calendari
+     * @param day dia del calendari
+     * @param method metode a executar
+     * @param object objecte del metode a executar
+     */
+    public void getDay(String userId, Date day, Method method, Object object)
+    {
+        CD.getDay(userId, day, method, object);
+    }
+
+    /**
+     * Retorna els dies de la base de dades del mes indicat
+     * @param userId id de l'usuari del calendari
+     * @param month més que volem seleccionar
+     * @param method metode a executar
+     * @param object objecte del metode a executar
+     */
+    public void getAvailableDays(String userId, int month, int year, Method method, Object object)
+    {
+        CD.getAvailableDays(userId, month, year, method, object);
+    }
+
+    /**
+     * Retorna els dies de la base de dades
+     * @param userId id de l'usuari del calendari
+     * @param method metode a executar
+     * @param object objecte del metode a executar
+     */
+    public void getAllDays(String userId, Method method, Object object)
+    {
+        CD.getAllDays(userId, method, object);
+    }
+
+    /**
+     * Retorna un enter amb el nombre de dies en ratxa que porta l'usuari
+     * @param userId id de l'usuari del calendari
+     * @param method metode a executar
+     * @param object objecte del metode a executar
+     */
+    public void getStreak(String userId, Method method, Object object)
+    {
+        CD.getStreak(userId,method,object);
+    }
+
+    /**
+     * Crea un dia al calendari nou.
+     * @param userId identificador de l'usuari
+     * @param day dia a crear.
+     * @param routineId id de la rutina a la que referencia
+     * @param totalActivities nombre d'activitats totals del dia de la rutina que estem afegint.
+     */
+    public String addDay(String userId, Date day, String routineId, int totalActivities)
+    {
+        return CD.addDay(userId, day,  routineId, totalActivities );
+    }
+
+    /**
+     * Actualitza la informació d'un dia
+     * @param userId id de l'usuari del calendari
+     * @param day dia del calendari
+     * @param activitiesDone nou nombre d'activitats fetes (a -1 no actualitzarà res)
+     * @param idRoutine nova id de la rutina a la que referencia
+     */
+    public void updateDay(String userId, Date day, int activitiesDone, String idRoutine) {
+        CD.updateDay(userId, day, activitiesDone, idRoutine);
+    }
+
+    /**
+     * Funcio que retorna els trofeus de l'usuari i un bool que indica per cada un si l'ha aconseguit
+     * @param userId identificador de l'usuari
+     * @param method metode que recull les dades
+     * @param object classe que conte el metode
+     */
+    public void getTrophies(String userId, Method method, Object object) {
+        CT.getTrophies(userId, method,object);
+    }
+
+    /**
+     * Funcio que afegeix un trofeu a l'usuari
+     * @param userId nom de l'usuari
+     * @param trophyName nom del trofeu
+     */
+    public void addTrophy(String userId, String trophyName) {
+        CT.addTrophy(userId, trophyName);
     }
 
 }
